v 2.0

+++ app.py
@@ -112,14 +112,14 @@ def update():
     _direccion=request.form['direccion']
     _tel=request.form['telForm']
     _estado=request.form['estado']
-    '''_foto=request.files['foto']'''
+    _foto=request.files['foto']
     id=request.form['idForm']
     '''now=datetime.now()
-    tiempo=now.strftime("%Y%H%M%S")'''
-    '''if _foto.filename !='':
+    tiempo=now.strftime("%Y%H%M%S")
+    if _foto.filename !='':
         nuevoNombreFoto=tiempo+_foto.filename
         _foto.save("uploads/"+nuevoNombreFoto)
-        cursor.execute("SELECT foto FROM `pacientesvet`.`pacientesvet` WHERE id=%s",(id))
+        cursor.execute("SELECT foto FROM pacientesvet WHERE id=%s",(id))
         fila= cursor.fetchall()
         os.remove(os.path.join(app.config['CARPETA'],fila[0][0]))
         cursor.execute("UPDATE pacientesvet SET foto=%s WHERE id=%s",(nuevoNombreFoto,id))
